Profile

+++ FlowZone.shared/Dtos/AvatarDto.cs
@@ -6,7 +6,7 @@ namespace FlowZone.shared.Dtos
     {
         public Guid AvatarId { get; set; }
         public string avatarName { get; set; } // Adjusted property name to match form input name
-        public double avatarPrice { get; set; } // Adjusted property name to match form input name
+        public int avatarPrice { get; set; } // Adjusted property name to match form input name
         public IFormFile avatarImage { get; set; }
         public string? ImagePath { get; set; }
     }
