+++ FlowZone.shared/Dtos/AvatarRequestDto.cs
@@ -10,7 +10,7 @@ namespace FlowZone.shared.Dtos
     public class AvatarRequestDto
     {
         public string avatarName { get; set; } // Adjusted property name to match form input name
-        public double avatarPrice { get; set; } // Adjusted property name to match form input name
+        public int avatarPrice { get; set; } // Adjusted property name to match form input name
         public IFormFile avatarImage { get; set; }
     }
 }
